Fix two compiler warnings.  No functional code changes.
FossilOrigin-Name: c829868aa2254c5e2268cdb803462cc6ec5cb71e

+++ src/util.c
@@ -1145,6 +1145,6 @@ int sqlite3MulInt64(i64 *pA, i64 iB){
 */
 int sqlite3AbsInt32(int x){
   if( x>=0 ) return x;
-  if( x==0x80000000 ) return 0x7fffffff;
+  if( x==(int)0x80000000 ) return 0x7fffffff;
   return -x;
 }
